refactor(ext/node): delete _fs_read.ts, port fs.readSync overload resolution to Rust
fs.readSync becomes a thin wrapper around op_node_fs_read_v_sync, which
replicates node's arity-based overload dispatch and exact validation order
(buffer first; fd validated last, and not at all when length resolves to 0).
fs.read moves into fs.ts as node-v26-verbatim JS (its dispatch allocates the
default buffer and locates the callback, and the zero-length case must
deliver via process.nextTick), now with FSReqCallback tracking and
callback-delivered node-formatted read errors via op_node_fs_read_deferred.

Also fixed to match the node v26 binary:
- position validated before the zero-length early-return (was after)
- error callback receives (err, 0, buffer) like node's wrapper (was (err, null))
- validatePosition type order is "bigint or integer"
- ERR_INVALID_ARG_TYPE renders bigints with the n suffix (JS validators);
  the binding-level fd check renders them without it, like node's C++
- ERR_OUT_OF_RANGE keeps decimal + _ separators up to 1e21 (was 9e15)
- proxied arrays are rejected as options bags (ArrayIsArray pierces proxies)

Enables test-fs-read-position-validation.mjs and
test-fs-read-promises-position-validation.mjs.

+++ ext/node/polyfills/internal/fs/utils.mjs
@@ -846,7 +846,8 @@ export const validatePosition = hideStackFrames((position, name, length) => {
       );
     }
   } else {
-    throw new ERR_INVALID_ARG_TYPE(name, ["integer", "bigint"], position);
+    // node v26's order ("bigint or integer", not "integer or bigint").
+    throw new ERR_INVALID_ARG_TYPE(name, ["bigint", "integer"], position);
   }
 });
